Installed and setup linting and formatting
Installed the tslint, the airbnb rules, prettier, and tslint-config-prettier to make the two work together.  Also setup config for both tslint and prettier and for tslint-config-prettier integration.

Finally, scripts were setup for fixing linting and formatting issues and for the corresponding ci commands such that these can be checked as part of the ci.