Setup Navigation
Setup the tab navigation that will be used in the app.  Also had to install several type def dependencies and change two lint rules.  variable-name was forcing a lowerCamelCase variable name, but it was going to be a component and needed to be Pascal case.  import-name was forcing me to import React as react, which would error on build.