Created modal to show images zoomed in
Created ImageModal to show images from the user’s list zoomed in full screen.  Navigation routes were adjusted.

Also, the import-name was turned off.  I don’t see the value in confusing names when it is written in the README for a library one way, I don’t want to go against that and this rule was fequently doing so.