Reworked Tabs to Stack
The idea of having three distinct tabs was just me trying to mock instagram.  I don’t think this app needs to, currently, be that complex.  So, the app is transitioning to have an Add Image button in the header and the User profile screen will be navigated to from a component on the UserImages screen.  Also, moved close button for modal to the right for ease of use.