User Account screen
The user account screen now shows a large user avatar that when pressed, allows the user to set their avatar and a text input that allows the user to adjust their username.  When either of these are set, the parent UserImages view is reloaded with data and the items are saved to disk.